Build function to generate new GWAS spreadsheet from pandas dataframe
Function attempts to deal with 3 of 5 possible output-format switches, and lays
groundwork for the reminaing 2 (which involve inserting extra columns into the
existing rows). Function runs but then appears to break, e.g. with
--format='in-situ'.

Adopt convention of using 'pd' as shorthand for pandas import.

Add idea for new switch to comments.