Use list of dictionaries for add operations: convert to DataFrame at end
Rather than attempting to write each tuple directly to the output dataframe, the
latter is converted (once only) from a list of dictionaries, with at least a
2-orders-of-magnitude performance improvement --- probably more, as performance
no longer degrades as the DataFrame grows.

Alternative means of growing the DataFrame tried but then commented out, as it
seems less performant than the original, and similarly suffers from progressive
degradation as the DataFrame grows.

There is no longer a hard limit on the number of records returned, or output.
This had been for testing purposes, but perhaps should be re-implemented as a
command switch, possibly returning a random subset.

Rename input tuple counter in anticipation of implementing output tuple counter:
decided to hold off on the latter, at least for this commit.

Straighten out mixed-up record set size reports, relating to rows in input
spreadsheet, and unique source terms.