Move active, non-boolean parameter defaults into external config file
All switch default values, bar inactive ones and booleans, ported to external
config. Includes the (currently sole) list parameter, 'target', which is
assigned its own config section. This allows ontology prefixes to be commented
and uncommented easily, minimising the chances of errors in specifying desired
targets (particularly as commented ontologies are likely to be greyed in most
editors); note that ontology prefixes are currently doubled up, as identical
key/value pairs: this is intentional but might be changed in the future.

Note that no consequences arise from not setting (superfluous) defaults on
options which, while permitted, which have no current effect.

Note further that because os.environ is imported into our ConfigParser, its
contents populate the 'DEFAULT' section of the config file, which then
propagates to every other section. This behaviour is unwanted, e.g. when using a
unique section to set values of list switches like 'target': we handle it by
turning the key/value pairs from both the 'DEFAULT' and 'Targets' sections into
sets, and subtracting the former from the latter.